Add sortable tables, fix movement parse and numeric types
Normalize movement detection to use viaje.direccion (rename ObtenerMovimientoDesdeRuta), preventing accidental normalization and null-silent rows. Convert several voyage/allocation counts from int to decimal (AsignacionExpedicionDto, SlideMezclaExpedicionDto, NodoComparativo armado dictionaries) and update accumulation logic to sum contribucion.Viajes. Introduce optional comparators and sorting UI/logic across multiple summary pages (ResumenAgenciasDesaparecidas, ResumenArbolComparativo, ResumenDestinosCayendo, ResumenRotacionOperadores, TablaOperadores) with visual arrows and table footers. Scope CSS rules for components (.re-page/.vz-page ::deep) and tweak minor UI text (Consulta button). Files changed: multiple files under Testing.Application.GetAllViajes and Testing.Components.Pages.

+++ Testing.Application/GetAllViajes/PresentacionModels.cs
@@ -50,10 +50,10 @@ public sealed record SlidePorClienteDto(IReadOnlyList<FilaClientePresentacionDto
 }
 
 public sealed record SlideMezclaExpedicionDto(
-    int Comodato,
-    int Full,
-    int Sencillo,
-    int Total,
+    decimal Comodato,
+    decimal Full,
+    decimal Sencillo,
+    decimal Total,
     decimal? PctComodato,
     decimal? DeltaPuntosPorcentuales,
     decimal VentaPorViajeFull,
